network: add ipv4ll as fallback for acd
When an address conflict is detected for the static IPv4 configuration a free
IPv4LL address has to be found. This patch adds functions start_ipv4ll for
starting the IPv4LL process, start_ipv4ll_ontimeout for a delayed start of
IPv4LL and remove_ipv4ll_timeout for stopping the delayed start.

For IPv4LL process again we use the new ACD functions.