network: make IPv4LL address for ACD persistent
After the occurance of an address conflict the fallback mechanism is to probe
for an IPv4LL address. According to RFC3927 (IPv4LL) after reboot the same
IPv4LL address should be used for the initial probing. This patch adds reading
of the IPv4 address previously used and if it was an IPv4LL address re-using
it after an address conflict.